LPS-60972 Isolate gradle-plugins-js-module-config-generator configuration
